+++ CHANGELOG.md
@@ -17,6 +17,7 @@ New features / improvements:
 - URLs can now span multiple lines, place '\\' at the end of the URL line to continue it in the next one. Leading whitespace in the additional lines will be skipped.
 - Added new function `verb-body-lf-to-crlf` designed for use with requests sending multipart data.
 - Added new `Verb-Prelude` heading property, which can be used to specify an Emacs Lisp or JSON file to load variables from, before performing requests.
+- Added `verb-shell` and `verb-unix-epoch` utility functions.
 
 ## **2.16.0** - 2024-03-02 (MELPA Stable)
 - Fixed LF being used instead of CRLF in multipart boundaries.